Create tic-tac-toe.py
This is a two player tic-tac-toe game which can be played in command-line. First, an empty game board is displayed and then the inputs are taken from the players. Secondly, it will check for the winning conditions if the condition satisfies it displays who is winner and if the condition is false and the whole board gets filled it displays draw.